requirements.txt
torch==1.7.1
tqdm==4.60.0
numpy==1.19.5
pandas==1.1.5
PyYAML==5.4.1
scikit-learn==0.24.2
